servidor/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-RESTX==1.3.0
SQLAlchemy==2.0.23
pyodbc==5.0.1
bcrypt==4.1.2
Pillow==10.1.0
python-dotenv==1.0.0
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
